add check config

+++ commands/clear-cache.js
@@ -1,13 +1,16 @@
 const exec = require("../utils/exec");
 const config = require("../config.json");
+const checkConfig = require("../utils/check-config");
 
 module.exports = function(program){
     program
     .command('clear-cache')
     .description('Effacer tous les caches des assets')
     .action(() => {
-        exec({
-            cmd: `cd ${config.projectpath}pixelhumain/ph/assets && rm -r ./*`
-        });
+        if(checkConfig()){
+            exec({
+                cmd: `cd ${config.projectpath}pixelhumain/ph/assets && rm -r ./*`
+            });
+        }
     });
 }
